Trait amethyst_renderer::pipe::pass::Pass

pub trait Pass: for<'a> PassData<'a> {
    fn compile(&mut self, effect: NewEffect) -> Result<Effect>;

    fn apply<'a, 'b: 'a>(
        &'a mut self, 
        encoder: &mut Encoder, 
        effect: &mut Effect, 
        factory: Factory, 
        data: Self::Data
    );
}

Structures implementing this provide a renderer pass.

Required methods

fn compile(&mut self, effect: NewEffect) -> Result<Effect>

The pass is given an opportunity to compile shaders and store them in an Effect which is then passed to the pass in apply.

fn apply<'a, 'b: 'a>(
&'a mut self,
encoder: &mut Encoder,
effect: &mut Effect,
factory: Factory,
data: Self::Data
)

Called whenever the renderer is ready to apply the pass. Feed commands into the encoder here.

Loading content...