Struct amethyst_renderer::pipe::EffectBuilder

pub struct EffectBuilder<'a> { /* fields omitted */ }

Methods

impl<'a> EffectBuilder<'a>

pub fn with_raw_global(&mut self, name: &'a str) -> &mut Self

Adds a global constant to this Effect.

pub fn with_raw_constant_buffer(
&mut self,
name: &'a str,
size: usize,
num: usize
) -> &mut Self

Adds a raw uniform constant to this Effect.

Requests a new constant buffer to be created

pub fn with_primitive_type(&mut self, prim: Primitive) -> &mut Self

Set the pipeline primitive type.

pub fn with_output(
&mut self,
name: &'a str,
depth: Option<DepthMode>
) -> &mut Self

Sets the output target of the PSO.

If the target contains a depth buffer, its mode will be set by depth.

pub fn with_blended_output(
&mut self,
name: &'a str,
mask: ColorMask,
blend: Blend,
depth: Option<DepthMode>
) -> &mut Self

Sets the output target of the PSO.

If the target contains a depth buffer, its mode will be set by depth.

pub fn with_texture(&mut self, name: &'a str) -> &mut Self

Adds a texture sampler to this Effect.

pub fn with_raw_vertex_buffer(
&mut self,
attrs: Attributes<'a>,
stride: ElemStride,
rate: InstanceRate
) -> &mut Self

Adds a vertex buffer to this Effect.

pub fn build(&mut self) -> Result<Effect>

TODO: Support render targets as inputs.