Struct amethyst_renderer::pipe::Stage

pub struct Stage<L> { /* fields omitted */ }

A stage in the rendering pipeline.

Methods

impl Stage<List<()>>

pub fn with_target<N: Into<String>>(target_name: N) -> StageBuilder<Queue<()>>

Builds a new PolyStage which outputs to the Target with the given name.

pub fn with_backbuffer() -> StageBuilder<Queue<()>>

Builds a new PolyStage which outputs straight into the backbuffer.

impl<L> Stage<L>

pub fn enable(&mut self)

Enables the PolyStage so it will execute on every frame.

pub fn disable(&mut self)

Disables the PolyStage, preventing it from being executed on every frame.

pub fn is_enabled(&self) -> bool

Returns whether this PolyStage is enabled.