Struct amethyst_renderer::pipe::Target

pub struct Target { /* fields omitted */ }

A render target.

Each render target contains a certain number of color buffers and an optional depth buffer.

Methods

impl Target

pub fn named<N: Into<String>>(name: N) -> TargetBuilder

Creates a new TargetBuilder with the given name.

pub fn clear_color<V: Into<[f32; 4]>>(&self, enc: &mut Encoder, value: V)

Clears all color buffers to the given value.

pub fn clear_depth_stencil<V: Into<f32>>(&self, enc: &mut Encoder, value: V)

Clears the depth stencil buffer to the given value.

pub fn size(&self) -> (u32, u32)

Returns the width and height of the render target, measured in pixels.

pub fn color_buf(&self, i: usize) -> Option<&ColorBuffer>

Returns the color buffer with index i.

pub fn color_bufs(&self) -> &[ColorBuffer]

Returns an array slice of the render target's color buffers.

pub fn depth_buf(&self) -> Option<&DepthBuffer>

Returns the render target's depth-stencil buffer, if it has one.