Module amethyst_renderer::prelude

Contains common types that can be glob-imported (*) for convenience.

Reexports

pub use input;

pub use light::*;

pub use mesh::Mesh;

pub use mesh::MeshBuilder;

pub use mesh::MeshHandle;

pub use pass;

pub use resources::*;

pub use system::RenderSystem;

pub use vertex::Color;

pub use vertex::Normal;

pub use vertex::PosColor;

pub use vertex::PosNormTangTex;

pub use vertex::PosNormTex;

pub use vertex::PosTex;

pub use vertex::Position;

pub use vertex::Separate;

pub use vertex::Tangent;

pub use vertex::TexCoord;

pub use vertex::VertexBufferCombination;

pub use vertex::VertexFormat;

Structs

Camera
Camera struct.

Material
Material struct.

MaterialDefaults
A resource providing default textures for Material. These will be be used by the renderer in case a texture handle points to a texture which is not loaded already. Additionally, you can use it to fill up the fields of Material you don't want to specify.

Pipeline
Defines how the rendering pipeline should be configured.

PipelineBuilder
Constructs a new pipeline with the given render targets and layers.

Renderer
Generic renderer.

Rgba
An RGBA color value.

Stage
A stage in the rendering pipeline.

StageBuilder
Constructs a new rendering stage.

Target
A render target.

Texture
Handle to a GPU texture resource.

TextureBuilder
Builds new textures.

TextureMetadata
Texture metadata, used while loading

Enums

MeshData
Mesh data for loading

Projection
The projection mode of a Camera.

TextureData
Texture data for loading

Traits

Pod
A trait for plain-old-data types.

PolyPipeline

PolyStage

Type Definitions

TextureHandle
A handle to a Texture asset.