Struct amethyst_renderer::ColorMask

pub struct ColorMask { /* fields omitted */ }

Methods

impl ColorMask

pub const RED: ColorMask

pub const GREEN: ColorMask

pub const BLUE: ColorMask

pub const ALPHA: ColorMask

pub const fn empty() -> ColorMask

Returns an empty set of flags

pub const fn all() -> ColorMask

Returns the set containing all flags.

pub const fn bits(&self) -> u8

Returns the raw value of the flags currently stored.

pub fn from_bits(bits: u8) -> Option<ColorMask>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u8) -> ColorMask

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects(&self, other: ColorMask) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: ColorMask) -> bool

Returns true all of the flags in other are contained within self.

pub fn insert(&mut self, other: ColorMask)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: ColorMask)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: ColorMask)

Toggles the specified flags in-place.

pub fn set(&mut self, other: ColorMask, value: bool)

Inserts or removes the specified flags depending on the passed value.