Struct amethyst_renderer::DebugLines

pub struct DebugLines {
    pub lines: Vec<PosColorNorm>,
}

Resource that stores non-persistent debug lines to be rendered in DebugLinesPass draw pass. The vector is automatically cleared after being rendered.

Fields
lines: Vec<PosColorNorm>

Lines to be rendered

Methods

impl DebugLines

pub fn new() -> DebugLines

Creates a new debug lines component with an empty DebugLine vector.

pub fn with_capacity(self, capacity: usize) -> Self

Builder method to pre-allocate a number of lines.

pub fn draw_direction(
&mut self,
position: Point3<f32>,
direction: Vector3<f32>,
color: Rgba
)

Submits a line to be rendered by giving a position and a direction.

pub fn draw_line(&mut self, start: Point3<f32>, end: Point3<f32>, color: Rgba)

Submits a line to be rendered by giving a start and an end position.