Struct amethyst_renderer::Layer

pub struct Layer {
    pub target: String,
    pub passes: Vec<Box<PassDescription>>,
}

A layer is comprised of a Render target and a list of passes

Fields
target: String

The render target, looked up by name during the Frame submission.

passes: Vec<Box<PassDescription>>

A list of passes to be executed in order to build up the target with the scene's data.

Methods

impl Layer

fn new<A>(target: A, passes: Vec<Box<PassDescription>>) -> Layer where String: From<A>

Create a new pass with that will target the supplied Target reference, The Layer will be initialized with the suppled list of passes.