Struct amethyst_renderer::MaterialPrefab

pub struct MaterialPrefab<F> where
    F: Format<Texture, Options = TextureMetadata>,  {
    pub albedo: Option<TexturePrefab<F>>,
    pub albedo_offset: TextureOffset,
    pub albedo_id: Option<u64>,
    pub emission: Option<TexturePrefab<F>>,
    pub emission_offset: TextureOffset,
    pub emission_id: Option<u64>,
    pub normal: Option<TexturePrefab<F>>,
    pub normal_offset: TextureOffset,
    pub normal_id: Option<u64>,
    pub metallic: Option<TexturePrefab<F>>,
    pub metallic_offset: TextureOffset,
    pub metallic_id: Option<u64>,
    pub roughness: Option<TexturePrefab<F>>,
    pub roughness_offset: TextureOffset,
    pub roughness_id: Option<u64>,
    pub ambient_occlusion: Option<TexturePrefab<F>>,
    pub ambient_occlusion_offset: TextureOffset,
    pub ambient_occlusion_id: Option<u64>,
    pub caveat: Option<TexturePrefab<F>>,
    pub caveat_offset: TextureOffset,
    pub caveat_id: Option<u64>,
    pub transparent: bool,
    pub alpha_cutoff: f32,
}

PrefabData for loading Materials

Type parameters:

F: Format to use for loading Textures

Fields
albedo: Option<TexturePrefab<F>>

Diffuse map.

albedo_offset: TextureOffset

Diffuse texture offset

albedo_id: Option<u64>

Albedo texture animation index, for addition to MaterialTextureSet

emission: Option<TexturePrefab<F>>

Emission map.

emission_offset: TextureOffset

Emission texture offset

emission_id: Option<u64>

Emission texture animation index, for addition to MaterialTextureSet

normal: Option<TexturePrefab<F>>

Normal map.

normal_offset: TextureOffset

Normal texture offset

normal_id: Option<u64>

Normal texture animation index, for addition to MaterialTextureSet

metallic: Option<TexturePrefab<F>>

Metallic map.

metallic_offset: TextureOffset

Metallic texture offset

metallic_id: Option<u64>

Metallic texture animation index, for addition to MaterialTextureSet

roughness: Option<TexturePrefab<F>>

Roughness map.

roughness_offset: TextureOffset

Roughness texture offset

roughness_id: Option<u64>

Roughness texture animation index, for addition to MaterialTextureSet

ambient_occlusion: Option<TexturePrefab<F>>

Ambient occlusion map.

ambient_occlusion_offset: TextureOffset

Ambient occlusion texture offset

ambient_occlusion_id: Option<u64>

Ambient occlusion texture animation index, for addition to MaterialTextureSet

caveat: Option<TexturePrefab<F>>

Caveat map.

caveat_offset: TextureOffset

Caveat texture offset

caveat_id: Option<u64>

Caveat texture animation index, for addition to MaterialTextureSet

transparent: bool

Set material as Transparent

alpha_cutoff: f32

Alpha cutoff: the value below which we do not draw the pixel