Struct amethyst_renderer::Mesh

pub struct Mesh { /* fields omitted */ }

Represents a polygonal mesh.

Methods

impl Mesh

pub fn build<D, V>(verts: D) -> MeshBuilder<((D, PhantomData<V>), ())> where
D: AsRef<[V]>,
V: VertexFormat,

Builds a new mesh from the given vertices.

pub fn buffer(&self, attributes: Attributes) -> Option<&RawBuffer<Resources>>

Returns the mesh's vertex buffer which matches requested attributes

pub fn slice(&self) -> &Slice<Resources>

Returns associated Slice

pub fn transform(&self) -> Matrix4<f32>

Returns the transformation matrix of the mesh.

This four-by-four matrix applies translation, rotation, and scaling to the mesh. It is often referred to in the computer graphics industry as the "model matrix".