Struct amethyst_renderer::MeshBuilder

pub struct MeshBuilder<D, V> { /* fields omitted */ }

Builds new meshes.

Methods

impl<D, V> MeshBuilder<D, V> where
D: AsRef<[V]>,
V: VertexFormat,

pub fn new(verts: D) -> Self

Creates a new MeshBuilder with the given vertices.

pub fn with_prim_type(self, prim: Primitive) -> Self

Sets the primitive type of the mesh.

By default, meshes are constructed as triangle lists.

pub fn with_position<P: Into<Point3<f32>>>(self, pos: P) -> Self

Sets the position of the mesh in 3D space.

pub fn with_rotation<Ax, An>(self, axis: Ax, angle: An) -> Self where
Ax: Into<Vector3<f32>>,
An: Into<Deg<f32>>,

Rotates the mesh a certain number of degrees around the given axis.

pub fn with_scale(self, val: f32) -> Self

Scales the mesh size according to the given value.

pub fn with_transform<M: Into<Matrix4<f32>>>(self, mat: M) -> Self

Sets the transformation matrix of the mesh.

This four-by-four matrix applies translation, rotation, and scaling to the mesh. It is often referred to in the computer graphics industry as the "model matrix".

pub fn build(self, fac: &mut Factory) -> Result<Mesh>

Builds and returns the new mesh.