Struct amethyst_renderer::RenderBundle

pub struct RenderBundle<'a, B, P> where
    B: PipelineBuild<Pipeline = P>,
    P: PolyPipeline,  { /* fields omitted */ }

Rendering bundle

Will register all necessary components needed for rendering, along with any resources. Will also register asset contexts with the asset Loader, and add systems for merging AssetFuture into its related component.

Will register TransparentSortingSystem, with name transparent_sorting_system if sorting is requested.

Methods

impl<'a, B, P> RenderBundle<'a, B, P> where
B: PipelineBuild<Pipeline = P>,
P: PolyPipeline,

pub fn new(pipe: B, config: Option<DisplayConfig>) -> Self

Create a new render bundle

pub fn with_visibility_sorting(self, dep: &'a [&'a str]) -> Self

Enable transparent mesh sorting, with the given dependencies

pub fn with_sprite_visibility_sorting(self, dep: &'a [&'a str]) -> Self

Enable transparent sprite sorting, with the given dependencies

pub fn with_sprite_sheet_processor(self) -> Self

Enable the sprite sheet processor

If you load a SpriteSheet in memory as an asset Format, this adds the Processor that will convert it to the Asset.