Struct amethyst_renderer::Renderer

pub struct Renderer<R: Resources, C: CommandBuffer<R>> {
    // some fields omitted
}

A Renderer manages passes and the execution of the passes over the targets. It only contains the passes, all other data is contained in the Frame

Methods

impl<R, C> Renderer<R, C> where R: Resources, C: CommandBuffer<R>

fn new(combuf: C) -> Renderer<R, C>

Create a new Render pipline

fn load_all<F>(&mut self, factory: &mut F) where F: Factory<R>

Load all known passes

fn add_pass<A, T, P>(&mut self, p: P) where P: Pass<R, Arg=A, Target=T> + 'static, A: PassDescription, T: Target

Add a pass to the table of available passes

fn submit<D>(&mut self, frame: &Frame<R>, device: &mut D) where D: Device<Resources=R, CommandBuffer=C>

Execute all passes