Struct amethyst_renderer::Scene

pub struct Scene { /* fields omitted */ }

Collection of lights and meshes to render.

Methods

impl Scene

pub fn add_light<L: Into<Light>>(&mut self, light: L)

Adds a light source to the scene.

Example

let mut scene = Scene::default();
scene.add_light(PointLight::default());

pub fn add_model(&mut self, model: Model)

Adds a mesh to the scene.

pub fn add_camera<C: Into<Camera>>(&mut self, camera: C)

Adds a camera to the scene.

pub fn set_ambient_color(&mut self, color: Rgba)

Set ambient color for the scene

pub fn ambient_color(&self) -> Rgba

Get the ambient color for the scene

pub fn lights(&self) -> &[Light]

Get all lights on scene

pub fn par_iter_lights(&self) -> Iter<'l, Light>

Iterates through all stored lights in parallel.

pub fn par_chunks_lights(&self, count: usize) -> Chunks<'l, Light>

Iterates through all stored lights in parallel in chunks.

pub fn par_iter_models(&self) -> Iter<'l, Model>

Iterates through all stored models in parallel.

pub fn par_chunks_models(&self, count: usize) -> Chunks<'l, Model>

Iterates through all stored models in parallel in chunks.

pub fn active_camera(&self) -> Option<&Camera>

Returns the active camera in the scene.

TODO: Render to multiple viewports with possibly different cameras.

pub fn clear(&mut self)

Remove all objects from Scene