Struct amethyst_renderer::ScreenDimensions

pub struct ScreenDimensions { /* fields omitted */ }

World resource that stores screen dimensions.

Methods

impl ScreenDimensions

pub fn new(w: u32, h: u32, hidpi: f64) -> Self

Creates a new screen dimensions object with the given width and height.

pub fn width(&self) -> f32

Returns the current width of the window.

pub fn height(&self) -> f32

Returns the current height of the window.

pub fn aspect_ratio(&self) -> f32

Returns the current aspect ratio of the window.

pub fn hidpi_factor(&self) -> f64

Returns the ratio between the backing framebuffer resolution and the window size in screen pixels. This is typically one for a normal display and two for a retina display.

pub fn update(&mut self, w: f64, h: f64)

Updates the width and height of the screen and recomputes the aspect ratio.

Only use this if you need to programmatically set the resolution of your game. This resource is updated automatically by the engine when a resize occurs so you don't need this unless you want to resize the game window.

pub fn update_hidpi_factor(&mut self, factor: f64)

Updates the hidpi factor stored in this structure.

Amethyst will call this for you automatically, most engine users won't need this.