Struct amethyst_renderer::Texture

pub struct Texture { /* fields omitted */ }

Handle to a GPU texture resource.

Methods

impl Texture

pub fn from_data<T: Pod + Copy, D: AsRef<[T]>>(data: D) -> TextureBuilder<D, T>

Builds a new texture with the given raw texture data.

pub fn from_color_val<C: Into<[f32; 4]>>(rgba: C) -> TextureBuilder<[u8; 4], u8>

Builds a new texture with the given raw texture data.

pub fn sampler(&self) -> &Sampler<Resources>

Returns the sampler for the texture.

pub fn view(&self) -> &RawShaderResourceView<Resources>

Returns the texture's raw shader resource view.