Struct amethyst_renderer::TextureMetadata

pub struct TextureMetadata {
    pub sampler: SamplerInfo,
    pub mip_levels: u8,
    pub dynamic: bool,
    pub format: SurfaceType,
    pub size: Option<(u16, u16)>,
    pub channel: ChannelType,
}

Additional texture metadata that can be passed to the asset loader or added to the prefab.

Fields
sampler: SamplerInfo

The sampler info describes how to read from the texture, thus specifies filter and wrap mode. The default is nearest filtering (FilterMethod::Scale) and clamping (WrapMode::Clamp).

mip_levels: u8

Mipmapping levels. The default is one level.

dynamic: bool

Dynamic texture

format: SurfaceType

The surface type of the texture which describes the number of color channels and their size. In simpler words, this defines the color format, e.g. RGBA 32-bit.

The default is R8_G8_B8_A8.

size: Option<(u16, u16)>

The dimensions of the texture. Only needed for raw image data (TextureData::U8 etc).

channel: ChannelType

The channel type which describes the data format of the channels (e.g. how the red value is stored).

This is usually Srgb for color textures, normalmaps & similar mostly use Unorm (which represents a value between 0.0 and 1.0).

Methods

impl TextureMetadata

pub fn unorm() -> Self

Creates texture metadata with Unorm channel type. This is used for normal / displacement maps. For color textures you most likely want to use TextureMetadata::srgb.

pub fn srgb() -> Self

Creates texture metadata for Srgb textures. This is usually the case for color textures.

For the values of all the other fields please refer to the documentation of the respective field.

pub fn srgb_scale() -> Self

Creates texture metadata for Srgb texture but without texture filtering. This is usually the case for color textures for 2D sprites and pixel art textures.

For the values of all the other fields please refer to the documentation of the respective field.

Wrap mode is set to WrapMode::Clamp by default.

pub fn with_sampler(self, info: SamplerInfo) -> Self

Sampler info

pub fn with_filter(self, filter: FilterMethod) -> Self

Sets the filter method of the sampler.

pub fn with_mip_levels(self, mip_levels: u8) -> Self

Mipmapping

pub fn with_format(self, format: SurfaceType) -> Self

Surface type

pub fn with_channel(self, channel: ChannelType) -> Self

Channel type

pub fn with_size(self, width: u16, height: u16) -> Self

Texture size

pub fn dynamic(self, d: bool) -> Self

Texture is dynamic