Module amethyst_renderer::target

Contains the included Render Targets

Structs

ColorBuffer
A simple output containing both a Color and a Depth target

GeometryBuffer
A geometry buffer is used in a deferred pipeline

Traits

Target
A Target or a RenderTarget is any object that can be the target of a Layer This is normally a framebuffer

Type Definitions

ColorFormat
Placeholder Color format

DepthFormat
Placeholder Depth Format