Trait amethyst_renderer::VertexFormat

pub trait VertexFormat: Pod + Sized + Send + Sync {
    const ATTRIBUTES: Attributes<'static>;
    fn size() -> usize { ... }

    fn attribute<F>() -> AttributeFormat
    where
        F: Attribute,
        Self: With<F>,
    { ... }
}

Trait implemented by all valid vertex formats.

Associated Constants

const ATTRIBUTES: Attributes<'static>

List of all attributes formats with name and offset.

Loading content...
Provided methods

fn size() -> usize

Returns the size of a single vertex in bytes.

fn attribute<F>() -> AttributeFormat where
F: Attribute,
Self: With<F>,

Returns attribute of vertex by type

Loading content...