Trait amethyst_renderer::With

pub trait With<F: Attribute>: VertexFormat {
    const FORMAT: AttributeFormat;
}

Trait implemented by all valid vertex formats for each field

Associated Constants

const FORMAT: AttributeFormat

Individual format of the attribute for this vertex format

Loading content...