Trait amethyst_renderer::vertex::VertexFormat

pub trait VertexFormat: Pod + Sized + Send + Sync {
    const ATTRIBUTES: &'static [(&'static str, AttributeFormat)];
    fn size() -> usize { ... }

    fn attribute<F>() -> AttributeFormat
    where
        Self: With<F>,
    { ... }
}

Trait implemented by all valid vertex formats.

Associated Constants

const ATTRIBUTES: &'static [(&'static str, AttributeFormat)]

List of all attributes formats with name and offset.

Provided Methods

fn size() -> usize

Returns the size of a single vertex in bytes.

fn attribute<F>() -> AttributeFormat where
Self: With<F>,

Returns attribute of vertex by type