Trait amethyst_renderer::vertex::WithField

pub trait WithField<F>: VertexFormat {
    fn field_attribute() -> Attribute;
}

Trait implemented by all valid vertex formats for each field

Required Methods

fn field_attribute() -> Attribute

Query individual attribute of the field for this format