Crate amethyst_renderer

A data parallel rendering engine developed by the Amethyst project. The source code is available for download on GitHub. See the online book for a complete guide to using Amethyst.

Reexports

pub use light::Light;

pub use vertex::VertexFormat;

Modules

light
Light sources.

pass
Different kinds of render passes.

pipe
Renderer pipeline configuration.

prelude
Contains common types that can be glob-imported (*) for convenience.

vertex
Built-in vertex formats.

Structs

Camera
Camera struct.

Config
Structure for holding the renderer configuration.

Material
Material struct.

MaterialBuilder
Builds new materials.

Mesh
Represents a polygonal mesh.

MeshBuilder
Builds new meshes.

PipelineBuilder
Constructs a new pipeline with the given render targets and layers.

Renderer
Generic renderer.

Rgba
An RGBA color value.

Target
A render target.

Texture
Handle to a GPU texture resource.

TextureBuilder
Builds new textures.

Enums

Error
Common renderer error type.

Projection
The projection mode of a Camera.

Traits

PolyPipeline

PolyStage

Type Definitions

Encoder
Command buffer encoder type.

Factory
Graphics factory type.

Result
Renderer result type.