Crate amethyst_renderer

High-level rendering engine with multiple backends.

Reexports

pub use pass::PassDescription;

pub use target::Target;

pub use pass::Pass;

Modules

pass
Contains the included Passes

target
Contains the included Render Targets

Structs

Camera
Contains the transforms for a Camera

ConstantColorTexture
holds a 1x1 texture that can be used to store constant colors

Fragment
A fragment is the most basic drawable element

Frame
The render job submission

Layer
A layer is comprised of a Render target and a list of passes

Light
A basic light

Renderer
A Renderer manages passes and the execution of the passes over the targets. It only contains the passes, all other data is contained in the Frame

Scene
A scene is a collection of fragments and lights that make up the scene.

VertexPosNormal

Enums

Texture