Struct amethyst_renderer::Renderer

pub struct Renderer {
    pub factory: Factory,
    // some fields omitted
}

Generic renderer.

Fields
factory: Factory

The gfx factory used for creation of buffers.

Methods

impl Renderer

pub fn new() -> Result<Renderer>

Creates a Renderer with default window settings.

pub fn build_with_loop(el: EventsLoop) -> RendererBuilder

Creates a new RendererBuilder, equivalent to RendererBuilder::new().

pub fn build() -> RendererBuilder

Creates a new RendererBuilder, equivalent to RendererBuilder::new().

pub fn create_mesh<T>(&mut self, mb: MeshBuilder<T>) -> Result<Mesh> where
T: VertexDataSet,

Builds a new mesh from the given vertices.

pub fn create_texture<D, T>(
&mut self,
tb: TextureBuilder<D, T>
) -> Result<Texture> where
D: AsRef<[T]>,
T: Pod + Copy,

Builds a new texture resource.

pub fn create_pipe<B, P>(&mut self, pb: B) -> Result<P> where
P: PolyPipeline,
B: PipelineBuild<Pipeline = P>,

Builds a new renderer pipeline.

pub fn draw<'a, P>(&mut self, pipe: &mut P, data: <P as PipelineData<'a>>::Data) where
P: PolyPipeline,

Draws a scene with the given pipeline.

pub fn events_mut(&mut self) -> &mut EventsLoop

Retrieve a mutable borrow of the events loop

pub fn resize<P: PolyPipeline>(&mut self, pipe: &mut P, new_size: (u32, u32))

Resize the targets associated with this renderer and pipeline.

pub fn window(&self) -> &WinitWindow

Retrieves an immutable borrow of the window.

No operations require a mutable borrow as of 2017-10-02