Struct amethyst_renderer::ScreenDimensions

pub struct ScreenDimensions { /* fields omitted */ }

World resource that stores screen dimensions.

Methods

impl ScreenDimensions

pub fn new(w: u32, h: u32, hidpi: f32) -> ScreenDimensions

Creates a new screen dimensions object with the given width and height.

pub fn width(&self) -> f32

Returns the current width of the window.

This is returned as a float for user convenience, as this is typically used with other float values. This will only ever be a non-negative integer though.

pub fn height(&self) -> f32

Returns the current height of the window.

This is returned as a float for user convenience, as this is typically used with other float values. This will only ever be a non-negative integer though.

pub fn aspect_ratio(&self) -> f32

Returns the current aspect ratio of the window.

pub fn hidpi_factor(&self) -> f32

Returns the ratio between the backing framebuffer resolution and the window size in screen pixels. This is typically one for a normal display and two for a retina display.

pub fn update(&mut self, w: u32, h: u32)

Updates the width and height of the screen and recomputes the aspect ratio.

Only use this if you need to programmatically set the resolution of your game. This resource is updated automatically by the engine when a resize occurs so you don't need this unless you want to resize the game window.