Struct amethyst_renderer::ScreenDimensions

pub struct ScreenDimensions { /* fields omitted */ }

World resource that stores screen dimensions.

Methods

impl ScreenDimensions

pub fn new(w: u32, h: u32) -> ScreenDimensions

Creates a new screen dimensions object with the given width and height.

pub fn width(&self) -> f32

Returns the current width of the window.

pub fn height(&self) -> f32

Returns the current height of the window.

pub fn aspect_ratio(&self) -> f32

Returns the current aspect ratio of the window.

pub fn update(&mut self, w: u32, h: u32)

Updates the width and height of the screen and recomputes the aspect ratio.