Trait amethyst_renderer::MeshCreator

pub trait MeshCreator: Send + Sync + Debug + 'static {
    fn build(self: Box<Self>, renderer: &mut Renderer) -> Result<Mesh>;
}

Trait used by the asset processor to convert any user supplied mesh representation into an actual Mesh.

This allows the user to create their own vertex attributes, and have the amethyst asset and render systems be able to convert it into a Mesh that can be used from any applicable pass.

Required Methods

fn build(self: Box<Self>, renderer: &mut Renderer) -> Result<Mesh>

Build a mesh given a Renderer